# Improve performance of URI#hostname(=) [RM#17219]