@nobu 2020-10-07
Have you tried `start_with?` and `end_with?`?.