@krzysiek1507 2020-10-07
I guess this patch should be submitted to the URI repo? https://github.com/ruby/uri