@lzap 2020-10-07
> Have you tried `start_with?` and `end_with?`?.

I haven't, I have considered it for a moment and thought that `[]` must be faster but this was indeed a C-language mindset! It's faster.

I also corrected a test failure (nil pointer exception missed).



> I guess this patch should be submitted to the URI repo? https://github.com/ruby/uri

Okay, apologies this is my first contribution. I tried to research contributing guides, looks like RedMine is preferred. I will update and close this PR, file new one as you recommend and also update the RM ticket with a patch. Thanks.