@krzysiek1507 2020-10-07
> Okay, apologies this is my first contribution.

No need to apologize. This repo is being synced with gem repos. Probably this patch should be submitted to the URI repo and then it will be synced here.