@lzap 2020-10-07
Okay, thanks. Moved to: https://github.com/ruby/uri/pull/12